strings_rmw.c:typecast char to int
Wasn't causing issues, but isspaces() takes an int for a value so I'm
assuming this might help prevent unforeseen problems?

+++ src/strings_rmw.c
@@ -106,7 +106,7 @@ trim (char *str)
   /* set pointer to segment preceding NULL terminator */
   str--;
 
-  while (isspace (*str))
+  while (isspace ((int)*str))
   {
     *str = '\0';
     str--;
